Update TEST_SPEED To Account For Negative Minimums (#153)
It's not uncommon for printers to have a negative `position_min`.  If you have a negative, the bounds of the Large moves may still be outside the bed moves. This is especially true if you have `position_min` incorrectly set at an excessively negative and incorrect amount.  This will cause the moves to still ram into the frame (...ask me how I know...).

This change forces it to use the lesser of the two values, zero or the `position_min`, causing it to be over the bed and preventing slams due to misconfiguration.